Version 1.0
I didn't commit the .ipch files because they were to large(over 100mb each).
The app will not run because is missing those files, If you want to run it send me a message and I will send them to you.
The files are found in: BankApp\.vs\BankApp\v15\ipch\AutoPCH